requirements.txt
panda
psycopg2
python-dotenv
psycop2-binary
sqlalchemy
matplotlib